(BKR-988) pull test tagging out of the structure module
